requirements/prod.txt
grpcio==1.69.0
grpcio-tools==1.69.0
python-dotenv==1.0.1

elasticsearch>=8.0.0
